Refactor Runner to be less tightly coupled
