+++ engine/launcher/src/main/scala/org/enso/launcher/components/runner/JVMSettings.scala
@@ -1,0 +1,3 @@
+package org.enso.launcher.components.runner
+
+case class JVMSettings(useSystemJVM: Boolean, jvmOptions: Seq[(String, String)])
