+++ engine/launcher/src/main/scala/org/enso/launcher/components/runner/RunSettings.scala
@@ -1,0 +1,5 @@
+package org.enso.launcher.components.runner
+
+import nl.gn0s1s.bump.SemVer
+
+case class RunSettings(version: SemVer, runnerArguments: Seq[String])
